+++ index.js
@@ -30,14 +30,48 @@ const convertFunctionEscapeForCSS = (rule) => {
  * @param root
  */
 const handleAtRules = (root) => {
-    root.walkAtRules(atRule => {
-        console.log('atrule = ', atRule);
+    const result = root.walkAtRules(atRule => {
+        console.log('AtRule = ', atRule);
+        // variables usage
         const variableRegex = /@/g;
         if (atRule.params.match(variableRegex)) {
             atRule.params = atRule.params.replace(variableRegex, '$');
         }
+        // convert file extension of @import At-Rules
+        if (atRule.name === 'import' && !atRule.params.match(/\.css/) && atRule.params.match(/['"]([a-z0-9-_]+)(\.\S+)?['"]/)) {
+            atRule.params = atRule.params.replace(/['"]([a-z0-9-_]+)(\.\S+)?['"]/, '"$1"')
+        }
     });
-    // variables reference
+    if (!result) {
+        console.log('walkAtRules failed');
+    }
+};
+
+const handleAtRule = (atRule) => {
+    console.log('AtRule = ', atRule);
+    // variables usage
+    const variableRegex = /@/g;
+    if (atRule.params.match(variableRegex)) {
+        atRule.params = atRule.params.replace(variableRegex, '$');
+    }
+    // // convert file extension of @import At-Rules
+    // if (atRule.name === 'import' && !atRule.params.match(/\.css/) && atRule.params.match(/['"]([a-z0-9-_]+)(\.\S+)?['"]/)) {
+    //     atRule.params = atRule.params.replace(/['"]([a-z0-9-_]+)(\.\S+)?['"]/, '"$1"');
+    // }
+};
+
+const handleImport = (importNode) => {
+    console.log('Import = ', importNode);
+    // convert file extension of @import At-Rules
+    if (!importNode.importPath.match(/\.css/)
+        && importNode.importPath.match(/['"]([a-z0-9-_\/]+)(\.\S+)?['"]/)) {
+        importNode.importPath = importNode.importPath.replace(/['"]([a-z0-9-_\/]+)(\.\S+)?['"]/, '"$1"');
+    }
+};
+
+const handleComment = (comment) => {
+    console.log('Comment = ', comment);
+    return comment;
 };
 
 /**
@@ -45,18 +79,45 @@ const handleAtRules = (root) => {
  * @param root
  */
 const handleDecls = (root) => {
-    root.walkDecls(decl => {
+    const result = root.walkDecls(decl => {
         console.log('decl.prop = ', decl.prop, ', decl.value = ', decl.value);
         convertFunctionSpin(decl);
         const variableRegex = /@/g;
+        // variable inside properties
         if (decl.prop.match(variableRegex)) {
             decl.prop = decl.prop.replace(variableRegex, '$');
         }
+        // variable inside values
         if (decl.value.match(variableRegex)) {
             decl.value = decl.value.replace(variableRegex, '$');
         }
+        // CSS escaping
+        decl.value = decl.value.replace(/~['"](.+)['"]/, '#{$1}')
     });
-    // variables reference
+    if (!result) {
+        console.log('walkDecls iteration was broke');
+    }
+};
+
+const handleDecl = (decl) => {
+    console.log('decl.prop = ', decl.prop, ', decl.value = ', decl.value);
+
+    convertFunctionSpin(decl);
+
+    // CSS escaping
+    decl.value = decl.value
+        .replace(/@{(\S+)}/, '$$$1')
+        .replace(/~['"](.+)['"]/, '#{$1}');
+
+    const variableRegex = /@/g;
+    // variable inside properties
+    if (decl.prop.match(variableRegex)) {
+        decl.prop = decl.prop.replace(variableRegex, '$');
+    }
+    // variable inside values
+    if (decl.value.match(variableRegex)) {
+        decl.value = decl.value.replace(variableRegex, '$');
+    }
 };
 
 /**
@@ -99,6 +160,39 @@ const handleRules = (root) => {
     // variables reference
 };
 
+const handleRule = (rule) => {
+    // mix-in properties from existing styles
+    const { selector, params, mixin } = rule;
+    console.log('rule.selector = ', selector,
+        ', rule.params = ', params,
+        ', rule.mixin = ', mixin);
+    // variables interpolation
+    if (selector.match(/\.@{(\S+)}/) && !selector.match(/~['"](.+)['"]/)) {
+        console.log('variable interpolation found, selector = ', selector);
+        rule.selector = selector.replace(/\.@{(\S+)}/g, '.#{$$$1}');
+    }
+    // mixin definition
+    if (selector.match(/\.\S+\((@\S+(:\s?\S+)?;?\s?)*\)/)) {
+        console.log('mixin definition found, selector = ', selector);
+        rule.selector = selector.replace(/@/g, '$')
+            .replace(/;/g, ',')
+            .replace('.', '@mixin ');
+    }
+    // mixin usage
+    if (mixin) {
+        console.log('mixin usage, found, selector = ', selector);
+        // rule = convertFunctionEscapeForCSS(rule);
+        rule.selector = selector
+            .replace(/~['"](.+)['"]/, '#{$1}')
+            .replace(/@{(\S+)}/, '$$$1')
+            .replace(/@/g, '$')
+            .replace(/;/g, ',')
+            .replace('.', '@include ')
+        ;
+        console.log('after mixin conversion, rule = ', rule);
+    }
+};
+
 module.exports = postcss.plugin('postcss-less2scss', function (opts) {
     opts = opts || {};
     console.log('opts = ', opts);
@@ -106,8 +200,36 @@ module.exports = postcss.plugin('postcss-less2scss', function (opts) {
 
     return function (root) {
         // Transform CSS AST here
-        handleAtRules(root);
-        handleDecls(root);
-        handleRules(root);
+        root.walk(node => {
+            const type = node.type;
+            console.log('node.type = ', type);
+
+            if (type === 'import') {
+                console.log('walk import');
+                handleImport(node);
+            }
+            if (type === 'root') {
+                console.log('walk root');
+            }
+            if (type === 'atrule') {
+                console.log('walk atrule');
+                handleAtRule(node);
+            }
+            if (type === 'rule') {
+                console.log('walk rule');
+                handleRule(node)
+            }
+            if (type === 'decl') {
+                console.log('walk decl');
+                handleDecl(node)
+            }
+            if (type === 'comment') {
+                console.log('walk comment');
+                handleComment(node)
+            }
+        })
+        // handleAtRules(root);
+        // handleDecls(root);
+        // handleRules(root);
     };
 });