+++ CHANGELOG.md
@@ -1,15 +1,23 @@
 # Change Log
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.5.2
+Bug fix:
+* Added CSS escaping conversion inside declaration values
+
+## 0.5.0
+Feature:
+* Convert file extension for @import at-rules
+
 ## 0.4.1
 Bug fix:
 * Support variables conversion of at-rule parameters
 
-## 0.4
+## 0.4.0
 Features:
 * Conversion for variable interpolation as selector name added
 
-## 0.3
+## 0.3.0
 Features:
 * Conversion for String Functions - CSS escaping, i.e., ~"" or ~'' added.
 
@@ -21,10 +29,10 @@ Bug fix
 Bug fix
 * Conversion for multiple variables usage fixed
 
-## 0.2
+## 0.2.0
 Features
 * Conversion for mixins and spin() color function added.
 
-## 0.1
+## 0.1.0
 Features
 * Conversion for variables definition and usage added.